+++ packages/playwright-core/src/server/chromium/crApp.ts
@@ -19,15 +19,6 @@ import path from 'path';
 import { isUnderTest } from '../../utils';
 import type { Page } from '../page';
 import { registryDirectory } from '../registry';
-import type { CRPage } from './crPage';
-
-export async function installAppIcon(page: Page) {
-  const icon = await fs.promises.readFile(require.resolve('./appIcon.png'));
-  const crPage = page._delegate as CRPage;
-  await crPage._mainFrameSession._client.send('Browser.setDockTile', {
-    image: icon.toString('base64')
-  });
-}
 
 export async function syncLocalStorageWithSettings(page: Page, appName: string) {
   if (isUnderTest())
